requirement.txt
# Core
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.0

# Database
SQLAlchemy==2.0.20

# Utilities
python-dotenv==1.0.0
bcrypt==4.0.1
pandas==2.0.3

# Testing & Development
pytest==7.4.0
black==23.7.0
flake8==6.0.0